pyproject.toml
[tool.poetry]
name = "decbot_web"
version = "0.1.0"
description = ""
authors = ["Peter Marheine <peter@taricorp.net>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.2.6"
djangorestframework = "^3.0"
django-filter = "<2.0"
django-pipeline = "^2.0.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
